pyproject.toml
[tool.poetry]
name = "nrdtech-python-aws-lambda"
version = "0.1.0"
description = ""
authors = ["Nic Delorme <nic.delorme@nrdtech.io>"]
license = "MIT License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.35.88"
uvicorn = "^0.34.0"
mangum = "^0.19.0"
python-dotenv = "^1.0.1"
requests = "2.32.5"
requests-oauthlib = "2.0.0"
logger = "1.4"
logger-config = "0.3"
psycopg2-binary = "2.9.11"
PyJWT = "^2.9.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
